MAPSJS-2660: BoundsGenerator cleanup.
- Move sphere view bounds and planar view bounds generation to separate
classes.
- Change interface to avoid having to change projection and wrapping
 both in MapView and BoundsGenerator.
